Adding bugs and homepage fields to package.json (#2401)
* Adding bugs and homepage fields to package.json

This adds bugs and homepage entries to the package.json. This helps these links render on npm, and can lead developers to the right place to put issues and to find more information. I have included the url which redirects to the github repository in case this changes later, for the homepage field.

Hopefully, this makes the package a bit more usable. Not a big fix, at all.

* Use https://github.com/zeit/next.js as the home